testing automated deployment (#2)
* minor update

* migrations for topics column used for streamlining contnet

* streamlining content set up

* testing streamlingin content and user creation with the same params

* lint fixes

* build docs

* upgrade packages

* automated deployment script
