+++ .gitignore
@@ -7,3 +7,4 @@ examples
 coverage/*
 cypress/videos
 cypress/screenshots
+TODO
