+++ appsec.yml
@@ -1,0 +1,17 @@
+version: 0.0
+os: linux
+files:
+  - source: /index.html
+    destination: /var/www/wikonnect/
+hooks:
+  BeforeInstall:
+    - location: scripts/install_dependencies
+      timeout: 300
+      runas: root
+    - location: scripts/start_server
+      timeout: 300
+      runas: root
+  ApplicationStop:
+    - location: scripts/stop_server
+      timeout: 300
+      runas: root
