+++ scripts/install_dependencies
@@ -1,0 +1,4 @@
+#!/bin/bash
+yarn
+yarn build:server
+yarn build:frontend
