+++ scripts/start_server
@@ -1,0 +1,2 @@
+#!/bin/bash
+pm2 start ecosystem.config.js
