+++ server/db/json_schema/userSchema.js
@@ -10,6 +10,7 @@ const userSchema = {
     'status': { 'type': 'string' },
     'profile_uri': { 'type': 'string' },
     'creatorId': { 'type': 'string' },
+    'topics': [{ 'type': 'string', }],
   },
   'required': ['username', 'hash', 'email'],
 };
