+++ server/db/seeds/chapters.js
@@ -17,7 +17,8 @@ exports.seed = function (knex) {
           content_uri: '/uploads/h5p/chapter1',
           created_at: '2017-12-20 19:17:10',
           updated_at: '2017-12-20 19:17:10',
-          approved: true
+          approved: true,
+          topics:  'primary'
         },
         {
           'id': 'chapter2',
@@ -30,7 +31,8 @@ exports.seed = function (knex) {
           'content_uri': '/uploads/h5p/chapter2',
           'created_at': '2017-12-20 19:17:10',
           'updated_at': '2017-12-20 19:17:10',
-          approved: true
+          approved: true,
+          topics: 'primary'
         },
         {
           'id': 'chapter3',
@@ -43,7 +45,8 @@ exports.seed = function (knex) {
           'content_uri': '/uploads/h5p/chapter3',
           'created_at': '2017-12-20 19:17:10',
           'updated_at': '2017-12-20 19:17:10',
-          approved: false
+          approved: true,
+          topics: 'highschool'
         },
         {
           'id': 'chapter4',
@@ -56,7 +59,8 @@ exports.seed = function (knex) {
           'content_uri': '/uploads/h5p/chapter4',
           'created_at': '2017-12-20 19:17:10',
           'updated_at': '2017-12-20 19:17:10',
-          approved: false
+          approved: true,
+          topics: 'university'
         }
       ]);
     });
