+++ server/db/seeds/users.js
@@ -17,6 +17,7 @@ exports.seed = function (knex) {
           last_seen: '2017-12-20 19:17:10',
           created_at: '2017-12-20 19:17:10',
           updated_at: '2017-12-20 19:17:10',
+          topics: ['highschool', 'primary', 'university'],
           metadata: { profileComplete: 'false', oneInviteComplete: 'false', oneChapterCompletion: 'false' }
         },
         {
@@ -30,6 +31,7 @@ exports.seed = function (knex) {
           last_seen: '2017-12-20 19:17:10',
           created_at: '2017-12-20 19:17:10',
           updated_at: '2017-12-20 19:17:10',
+          topics: ['highschool', 'primary'],
           metadata: { profileComplete: 'false', oneInviteComplete: 'false', oneChapterCompletion: 'false' }
         },
         {
@@ -43,6 +45,7 @@ exports.seed = function (knex) {
           last_seen: '2017-12-20 19:17:10',
           created_at: '2017-12-20 19:17:10',
           updated_at: '2017-12-20 19:17:10',
+          topics: ['highschool', 'university'],
           metadata: { profileComplete: 'true', oneInviteComplete: 'true', oneChapterCompletion: 'true' }
         }
       ]);
