+++ server/routes/chapters.js
@@ -11,6 +11,7 @@ const log = require('../utils/logger');
 
 
 const Chapter = require('../models/chapter');
+const User = require('../models/user');
 const permController = require('../middleware/permController');
 const validateChapter = require('../middleware/validation/validateChapter');
 
@@ -20,34 +21,10 @@ const router = new Router({
   prefix: '/chapters'
 });
 
-async function returnType(parent) {
-  if (parent.length == undefined) {
-    parent.comment.forEach(comment => {
-      return comment.type = 'comment';
-    });
-  } else {
-    parent.forEach(mod => {
-      mod.comment.forEach(comment => {
-        return comment.type = 'comment';
-      });
-    });
-  }
-}
-
-async function achievementType(parent) {
-  if (parent.length == undefined) {
-    parent.achievement.forEach(data => {
-      return data.type = 'achievement';
-    });
-  } else {
-    parent.forEach(mod => {
-      mod.achievement.forEach(data => {
-        return data.type = 'achievement';
-      });
-    });
-  }
-}
-
+const {
+  returnType,
+  achievementType
+} = require('../utils/routesUtils/chaptersRouteUtils');
 
 /**
  * @api {get} /chapters/ GET all chapters.
@@ -75,6 +52,7 @@ async function achievementType(parent) {
  *            "imageUrl": "/uploads/images/content/chapters/chapter1.jpeg",
  *            "contentId": null,
  *            "tags": [],
+ *            topic: ''
  *            "comment": [{
  *            }]
  *         }]
@@ -85,16 +63,18 @@ async function achievementType(parent) {
 router.get('/', permController.requireAuth, async ctx => {
 
   let stateUserRole = ctx.state.user.role == undefined ? ctx.state.user.data.role : ctx.state.user.role;
+  let stateUserId = ctx.state.user.id == undefined ? ctx.state.user.data.id : ctx.state.user.id;
+
   let roleNameList = ['basic', 'superadmin', 'tunapanda', 'admin'];
+  let user = await User.query().findById(stateUserId);
 
   let chapter;
   if (roleNameList.includes(stateUserRole)) {
     if (ctx.query.q) {
-      chapter = await Chapter.query()
-        .select('chapters.*')
-        .avg('rate.rating as rating')
-        .from('chapters')
+      chapter = await chapter
         .where('name', 'ILIKE', `%${ctx.query.q}%`)
+        .where(ctx.query, { status: 'published' })
+        .whereIn('topics', user.topics)
         .orWhere('description', 'ILIKE', `%${ctx.query.q}%`)
         .where({ status: 'published' })
         .leftJoin('ratings as rate', 'chapters.id', 'rate.chapter_id')
@@ -106,6 +86,7 @@ router.get('/', permController.requireAuth, async ctx => {
         .avg('rate.rating as rating')
         .from('chapters')
         .where(ctx.query, { status: 'published' })
+        .whereIn('topics', user.topics)
         .leftJoin('ratings as rate', 'chapters.id', 'rate.chapter_id')
         .groupBy('chapters.id', 'rate.chapter_id')
         .eager('[comment(selectComment), achievement(selectAchievement), flag(selectFlag)]');
@@ -125,7 +106,7 @@ router.get('/', permController.requireAuth, async ctx => {
   }
 
   ctx.status = 200;
-  ctx.body = { 'chapter': chapter };
+  ctx.body = { chapter };
 });
 
 router.get('/teach', permController.requireAuth, async ctx => {
@@ -184,6 +165,8 @@ router.get('/:id', permController.requireAuth, async ctx => {
 
   let roleNameList = ['basic', 'superadmin', 'tunapanda'];
   let anonymous = 'anonymous';
+  let user = await User.query().findById(stateUserId);
+  // let chapter = Chapter.query().where({ 'chapters.id': ctx.params.id, status: 'published' });
 
   let chapter = Chapter.query()
     .select('chapters.*')
@@ -194,13 +177,12 @@ router.get('/:id', permController.requireAuth, async ctx => {
     .groupBy('chapters.id', 'rate.chapter_id');
 
   if (roleNameList.includes(stateUserRole)) {
-    chapter = await chapter.eager('[comment(selectComment), flag(selectFlag), achievement(selectAchievement)]');
+    chapter = await chapter.whereIn('topics', user.topics).eager('[comment(selectComment), flag(selectFlag), achievement(selectAchievement)]');
     await achievementType(chapter);
   } else if (stateUserRole == anonymous) {
     chapter = await chapter.eager('comment(selectComment)');
   } else {
-    chapter = await Chapter.query()
-      .where({ id: ctx.params.id, creatorId: stateUserId });
+    chapter = await Chapter.query().where({ id: ctx.params.id, creatorId: stateUserId });
   }
 
   ctx.assert(chapter, 404, 'no lesson by that ID');
